Local Binary Pattern algorithm
Done a quick implementation of LBP following the method that uses
uniform LBP values and drives the nonuniform values to a single bin.

I created a small test script to play around with the function. Still
need to verify it using our Data base (still need to download it lol)