Cleaned up LBP3.m:
Stripped out a handful of unneccesary code.
The binning of those with more that 2 transitions didn't seem useful,
LPQ handles 256 features fine so it felt like we were just throwing
away information. Optimized the windowing calculation, the weights
matrix when multiplied gets rid of the center value. The slight
variation noticed earlier was due to the rotation starting at a
different position. The accuracy on the test data is improved to
nearly the same as LPQ.

Using LBP3 the training and test data is processed significantly faster.
~3 seconds as opposed to 34 seconds on the test file images.

Cleaned up the sliding window in LinearBinaryPattern.m. Cannot use a
parfor loop but isnt really neccesary. In the main script the function
is called by a parallel loop so I think it is about as good as it gets.